refactor: erc20 util function
