+++ .github/workflows/node.js.yml
@@ -29,6 +29,7 @@ jobs:
     - run: yarn install
     - run: yarn run build
     - run: yarn workspace @godwoken-web3/api-server run test tests/db/helpers.test.ts
+    - run: yarn workspace @godwoken-web3/api-server run test tests/utils/erc20.test.ts
     - run: yarn run fmt
     - run: yarn run lint
     - run: git diff --exit-code
