+++ packages/api-server/package.json
@@ -29,7 +29,6 @@
     "@ckb-lumos/toolkit": "0.18.0-rc6",
     "@godwoken-web3/godwoken": "1.0.0-rc1",
     "@newrelic/native-metrics": "^7.0.1",
-    "@polyjuice-provider/base": "^0.1.5",
     "@sentry/node": "^6.11.0",
     "blake2b": "2.1.3",
     "cors": "^2.8.5",
