+++ packages/api-server/tests/utils/erc20.test.ts
@@ -1,0 +1,16 @@
+import test from "ava";
+import { isErc20Transfer } from "../../src/erc20";
+
+test("isErc20Transfer", (t) => {
+  const emptyInput = "0x";
+  const illegalInput = "123";
+  const uniswapInput =
+    "0x0dcd7a6c000000000000000000000000f0d9e03b98be925886a5bb9e031920a5c17d333d0000000000000000000000000000000000000000000000000000000013e6c196000000000000000000000000dac17f958d2ee523a2206206994597c13d831ec7000000000000000000000000000000000000000000000000000000006267855000000000000000000000000000000000000000000000000000000000002085cc00000000000000000000000000000000000000000000000000000000000000c0000000000000000000000000000000000000000000000000000000000000004136aff2130612bbad08900110dd05f33609db363ea40b088df3a396ed1da8f4f724c912a2323ca94978d0b6bebc2cae6fe19bb7de7ff19c355394fa4f6900a48e1b00000000000000000000000000000000000000000000000000000000000000";
+  const erc20TransferInput =
+    "0xa9059cbb000000000000000000000000477b8d5ef7c2c42db84deb555419cd817c336b6f000000000000000000000000000000000000000000000000000000001ad27480";
+
+  t.false(isErc20Transfer(emptyInput));
+  t.false(isErc20Transfer(illegalInput));
+  t.false(isErc20Transfer(uniswapInput));
+  t.true(isErc20Transfer(erc20TransferInput));
+});
